+++ src/OpenTelemetry/Logs/ILogger/OpenTelemetryLoggerProvider.cs
@@ -15,9 +15,11 @@ namespace OpenTelemetry.Logs;
 [ProviderAlias("OpenTelemetry")]
 public class OpenTelemetryLoggerProvider : BaseProvider, ILoggerProvider, ISupportExternalScope
 {
-    internal readonly LoggerProvider Provider;
+    private readonly Lock syncObject = new();
     private readonly bool ownsProvider;
     private readonly Hashtable loggers = [];
+    private Func<LoggerProvider>? loggerProviderFactory;
+    private LoggerProvider? provider;
     private bool disposed;
 
     static OpenTelemetryLoggerProvider()
@@ -40,7 +42,7 @@ public class OpenTelemetryLoggerProvider : BaseProvider, ILoggerProvider, ISuppo
         var optionsInstance = options.CurrentValue;
 #pragma warning restore CA1062 // Validate arguments of public methods - needed for netstandard2.1
 
-        this.Provider = Sdk
+        this.provider = Sdk
             .CreateLoggerProviderBuilder()
             .ConfigureBuilder((sp, builder) =>
             {
@@ -61,13 +63,40 @@ public class OpenTelemetryLoggerProvider : BaseProvider, ILoggerProvider, ISuppo
     }
 
     internal OpenTelemetryLoggerProvider(
-        LoggerProvider loggerProvider,
+        Func<LoggerProvider> loggerProviderFactory,
         OpenTelemetryLoggerOptions options,
         bool disposeProvider)
     {
-        this.Provider = loggerProvider;
+        Guard.ThrowIfNull(loggerProviderFactory);
+
+        this.loggerProviderFactory = loggerProviderFactory;
         this.Options = options.Copy();
         this.ownsProvider = disposeProvider;
+    }
+
+    internal LoggerProvider Provider
+    {
+        get
+        {
+            var provider = this.provider;
+            if (provider != null)
+            {
+                return provider;
+            }
+
+            lock (this.syncObject)
+            {
+                provider = this.provider;
+                if (provider == null)
+                {
+                    provider = this.loggerProviderFactory!();
+                    this.provider = provider;
+                    this.loggerProviderFactory = null;
+                }
+
+                return provider;
+            }
+        }
     }
 
     internal OpenTelemetryLoggerOptions Options { get; }
@@ -130,7 +159,7 @@ public class OpenTelemetryLoggerProvider : BaseProvider, ILoggerProvider, ISuppo
             {
                 if (this.ownsProvider)
                 {
-                    this.Provider.Dispose();
+                    this.provider?.Dispose();
                 }
             }