+++ test/OpenTelemetry.Exporter.OpenTelemetryProtocol.Tests/OtlpLogExporterTests.cs
@@ -1728,6 +1728,10 @@ public class OtlpLogExporterTests
 
         Assert.NotNull(factory);
 
+        var logger = factory.CreateLogger("TestLogger");
+
+        Assert.NotNull(logger);
+
         Assert.True(configureDelegateCalled);
         Assert.True(configureExportProcessorOptionsCalled);
         Assert.True(configureBatchOptionsCalled);