+++ src/OpenTelemetry/CHANGELOG.md
@@ -22,6 +22,9 @@ Notes](../../RELEASENOTES.md).
 * Update `OpenTelemetrySdkEventSource` to support the W3C randomness flag.
   ([#7301](https://github.com/open-telemetry/opentelemetry-dotnet/pull/7301))
 
+* Fix circular reference when `LoggerProvider` requires `ILogger`.
+  ([#7308](https://github.com/open-telemetry/opentelemetry-dotnet/pull/7308))
+
 ## 1.15.3
 
 Released 2026-Apr-21